#222 internal dto have been moved from Long to Int

+++ app/src/main/java/io/golos/cyber_android/ui/common/formatters/counts/KiloCounterFormatter.kt
@@ -4,7 +4,7 @@ import java.util.*
 import kotlin.math.abs
 
 object KiloCounterFormatter {
-    fun format(count: Long): String =
+    fun format(count: Int): String =
         when {
             abs(count) < 1000 -> count.toString()
             abs(count) % 1000 < 51 -> "${count/1000}k"
