+++ app/src/main/java/io/golos/cyber_android/ui/common/formatters/size/SizeFormatterBase.kt
@@ -5,12 +5,12 @@ import kotlin.math.log10
 import kotlin.math.pow
 
 abstract class SizeFormatterBase(private val calculationBase: Double): SizeFormatter {
-    fun calculateUnitIndex(size: Long): Int =
+    fun calculateUnitIndex(size: Int): Int =
         if(size <= 0)
             0
         else
             (log10(size.toDouble()) / log10(calculationBase)).toInt()
 
-    fun calculateValue(template: String, size: Long, unitIndex: Int) =
+    fun calculateValue(template: String, size: Int, unitIndex: Int) =
         DecimalFormat(template).format(size/ calculationBase.pow(unitIndex))
 }
